BACKEND/requirements.txt
# List all the Python packages that are required for the project to run
Flask==3.0.2
Pillow==10.2.0
numpy==1.26.4
tensorflow==2.16.1
Werkzeug==3.0.1
